helssa-big-back-main/agent/sample_codes/minio-example/serializers.py
# serializers.py

from rest_framework import serializers
from .models import File

class FileSerializer(serializers.ModelSerializer):
    class Meta:
        model = File
        fields = ['id', 'file_type', 'file', 'created_at']
